requirements.txt
tweepy>=4.14.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pymongo>=4.6.0
detoxify>=0.5.2
pydantic>=2.5.0
pandas>=2.0.0
python-dotenv>=1.0.0